refactor: modify field type of id

+++ src/main/java/techcourse/myblog/domain/Article.java
@@ -3,7 +3,7 @@ package techcourse.myblog.domain;
 import java.util.Objects;
 
 public class Article {
-	private int id;
+	private long id;
 	private String title;
 	private String contents;
 	private String coverUrl;
@@ -38,15 +38,15 @@ public class Article {
 		this.contents = contents;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public boolean matchId(int articleId) {
+	public boolean matchId(long articleId) {
 		return this.id == articleId;
 	}
